More build errors and warnings fixed
5 errors remaining at this point.

Only warning left are in the Stormlib library.

+++ src/game/Object/Guild.cpp
@@ -162,7 +162,7 @@ void Guild::CreateDefaultGuildRanks(int locale_idx)
     CreateRank(sObjectMgr.GetMangosString(LANG_GUILD_MEMBER, locale_idx),   GR_RIGHT_GCHATLISTEN | GR_RIGHT_GCHATSPEAK);
     CreateRank(sObjectMgr.GetMangosString(LANG_GUILD_INITIATE, locale_idx), GR_RIGHT_GCHATLISTEN | GR_RIGHT_GCHATSPEAK);
 
-    SetBankMoneyPerDay((uint32)GR_GUILDMASTER, WITHDRAW_MONEY_UNLIMITED);
+    SetBankMoneyPerDay((uint32)GR_GUILDMASTER, (uint32)WITHDRAW_MONEY_UNLIMITED);
 }
 
 bool Guild::AddMember(ObjectGuid plGuid, uint32 plRank)
@@ -1561,7 +1561,7 @@ void Guild::SetBankMoneyPerDay(uint32 rankId, uint32 money)
         return;
 
     if (rankId == GR_GUILDMASTER)
-        money = WITHDRAW_MONEY_UNLIMITED;
+        money = (uint32)WITHDRAW_MONEY_UNLIMITED;
 
     m_Ranks[rankId].BankMoneyPerDay = money;
 
@@ -1611,7 +1611,7 @@ uint32 Guild::GetBankMoneyPerDay(uint32 rankId)
         return 0;
 
     if (rankId == GR_GUILDMASTER)
-        return WITHDRAW_MONEY_UNLIMITED;
+        return (uint32)WITHDRAW_MONEY_UNLIMITED;
     return m_Ranks[rankId].BankMoneyPerDay;
 }
